refactor: split  Dashboard integration tests to separate files
